Update lint report filename regex
In versions < 2.3.0 of the Android Gradle plugin, lint report output
was placed into the build folder's outputs directory, and was named
lint-results-[variant].xml. As of version 2.3.0, the plugin now places
lint reports into the build folder's reports directory, and names the
report file lint-results.xml. This commit updates the path used to
locate lint report files. It is fairly defensive and should be backwards
compatible. We do not yet fully understand whether a suffix will ever be
present. If multiple lint files are present, there are no guarantees
that Gnag will select a specific (or even the newest) file. If you
encounter an issue with this strategy, please file an issue detailing
your use-case, Gradle and plugin versions, Gradle invocations, etc.