Improve extension APIs for Kotlin users
Standardizes reporter enabled boolean accessor methods according to
https://kotlinlang.org/docs/reference/java-interop.html#getters-and-setters
so that they can be interacted with as Kotlin properties.

The old methods have been deprecated and should be removed in the next
major version of Gnag (since their removal will be a breaking API change
for Kotlin-based build scripts).

Groovy-based build scripts should not need to be updated in response to
this change.

Kotlin-based build scripts should not need to be updated in response to
this change.